OP-225: Refactor forms
